:truck: 최근 당첨 결과 확인 기능을 하나의 파일로 추출한다
