+++ main.py
@@ -1,12 +1,5 @@
-from datetime import timedelta
-from textwrap import dedent
-from typing import Any
-
 from lotto.account import Account, fetch_account
 from lotto.lotto import *
-from lotto.secret import Secret
-from sends.send import Send
-from sends.send_github_issue import SendGithubIssue
 
 
 class LottoError(Exception):
@@ -66,28 +59,6 @@ def check_lottery_result(start_date: date, end_date: date) -> dict[str, int]:
     return {'시작일': start_date, '종료일': end_date} | total_buy_result(buy_results())
 
 
-def to_message(result: dict[str, Any]) -> str:
-    return dedent(f'''\
-    💰 총 당첨금: {"{:,}".format(result["총 당첨금"])}원
-    ✅ 총 구입매수: {result["총 구입매수"]}장 (미추첨 {result["미추첨"]}장)
-    📅 조회기간: {result["시작일"].strftime("%y-%m-%d")} ~ {result["종료일"].strftime("%y-%m-%d")}''')
-
-
-def last_sunday(today: date) -> date:
-    days_in_week = 7
-    pass_days_in_last_week = len(['일요일'])
-    pass_days = (today.weekday() + pass_days_in_last_week) % days_in_week
-
-    return today - timedelta(days=pass_days)
-
-
 if __name__ == '__main__':
     login(fetch_account())
     # buy(amount=1)
-    lottery_result = check_lottery_result(start_date=last_sunday(date.today()), end_date=date.today())
-    send: Send = SendGithubIssue(
-        token=Secret('ghp_'),
-        repository='viiviii/jubilant-train'
-    )
-    send(title='🎊 로또6/45 1055회(23-02-18)',  # todo: title 하드 코딩 제거
-         content=to_message(lottery_result))
